+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 * Detection of battery type
 * Remove special characters from username to generate valid MQTT topics
 * Setting ha_discovery_enabled to False had no effect
+* Docker image based on python:3.11-slim
+* Force refresh by charging station only if charging value has changed
 
 ## 0.4.5